Rename runCommonViewTests to runCommonExtensionViewTests.
