+++ test/functional/runCommonExtensionViewTests.js
